Move Workout UI models to separate files
